Fixes en ventana de oferta y validacion

+++ database/migrations/2015_06_23_181917_create_ofertas_table.php
@@ -21,7 +21,7 @@ class CreateOfertasTable extends Migration {
 			$table->integer('user_id')->unsigned();
 			$table->foreign('user_id')->references('id')->on('users');
 			//--
-			$table->float('monto');
+			$table->integer('monto');
 			$table->longtext('necesidad');
 			$table->timestamps();
 		});
